Move helper methods back to Twitter4JService
